[Spyll] Move files around in Spyll, setup for including from other solutions
fixup
