+++ Source/Spyll/Tests/MultiTranslationUnit.cpp
@@ -1,4 +1,4 @@
-#include <Spyll/ClangTool/ReflectionGenerator.h>
+#include "Tool/Clang/ReflectionGenerator.h"
 
 #include "ToolPreTest.h"
