+++ Source/Spyll/Tool/Core/DescriptorEnums.h
@@ -2,32 +2,15 @@
 
 #include <type_traits>
 
-#include "Spyll/Spyll/Enums.h"
+#include "Tool/Core/Include/Enums.h"
 
 namespace Spyll
 {
-#define JSON_ENUM_HELPER_FNS(_name_) \
-	template<typename BasicJsonType> \
-	void \
-	to_json(BasicJsonType& a_json, const _name_& a_id) \
-	{ \
-		a_json = static_cast<std::underlying_type_t<_name_>>(a_id); \
-	} \
-	\
-	template<typename BasicJsonType> \
-	void \
-	from_json(const BasicJsonType& a_json, _name_& a_id) \
-	{ \
-		a_id = static_cast<_name_>(a_json.template get<std::underlying_type_t<_name_>>()); \
-	} \
-
 #define DESCRIPTOR_ID(_name_) \
 	enum class _name_ : uint32_t \
 	{ \
 		Invalid = static_cast<std::underlying_type_t<_name_>>(-1)\
 	}; \
-	\
-	JSON_ENUM_HELPER_FNS(_name_) \
 	\
 	inline \
 	bool \
@@ -89,16 +72,12 @@ namespace Spyll
 		return a_spec == Reflection::AccessSpecifier::None;
 	}
 
-	JSON_ENUM_HELPER_FNS(Reflection::AccessSpecifier)
-
 	inline
 	bool
 	operator !(Reflection::ConstructorType a_type)
 	{
 		return a_type == Reflection::ConstructorType::None;
 	}
-
-	JSON_ENUM_HELPER_FNS(Reflection::ConstructorType)
 }
 
 #undef JSON_ENUM_HELPER_FNS
