+++ Source/Spyll/Tool/Core/DescriptorSerialization.h
@@ -20,6 +20,21 @@
 	if (auto iter = _json_.find(#_name_); iter != _json_.end())\
 		iter->get_to(_obj_._name_)
 
+#define JSON_ENUM_HELPER_FNS(_name_) \
+	template<typename BasicJsonType> \
+	void \
+	to_json(BasicJsonType& a_json, const _name_& a_id) \
+	{ \
+		a_json = static_cast<std::underlying_type_t<_name_>>(a_id); \
+	} \
+	\
+	template<typename BasicJsonType> \
+	void \
+	from_json(const BasicJsonType& a_json, _name_& a_id) \
+	{ \
+		a_id = static_cast<_name_>(a_json.template get<std::underlying_type_t<_name_>>()); \
+	} \
+
 namespace Spyll
 {
 	// ReflectionDescriptor
@@ -69,4 +84,18 @@ namespace Spyll
 	void
 	from_json(const nlohmann::json& a_json, EnumDescriptor& a_descriptor);
 	// EnumDescriptor
+
+	JSON_ENUM_HELPER_FNS(TypeDescriptorId)
+
+	JSON_ENUM_HELPER_FNS(FieldDescriptorId)
+
+	JSON_ENUM_HELPER_FNS(FunctionDescriptorId)
+
+	JSON_ENUM_HELPER_FNS(VariableDescriptorId)
+
+	JSON_ENUM_HELPER_FNS(EnumDescriptorId)
+
+	JSON_ENUM_HELPER_FNS(Reflection::AccessSpecifier)
+
+	JSON_ENUM_HELPER_FNS(Reflection::ConstructorType)
 }
