+++ Source/Spyll/Tool/Core/EnumDescriptor.cpp
