+++ Source/Spyll/Tool/Core/Enums.h
