+++ Source/Spyll/Tool/Core/FunctionDescriptor.cpp
