+++ Source/Spyll/Tool/Core/ReflectionDescriptor.cpp
