+++ Source/Spyll/Tool/Core/ReflectionDescriptor.h
