+++ Source/Spyll/Tool/Core/Spyll.Core.natvis
