+++ Source/Spyll/Tool/Core/TypeDescriptor.cpp
