+++ Source/Spyll/Tool/Core/TypeDescriptor.h
