+++ Source/Spyll/Tool/Core/VariableDescriptor.cpp
