+++ Source/Spyll/Tool/Core/VariableDescriptor.h
